Create README.md
This application should consist of two WPF applications. First server, second - Client. On one computer, the server must be started, the client runs on the workstations. The client should be able to enter his name and select the color that the user name will be displayed in the chat. When the user clicks on the "Join" button, a window opens that displays a list of users in the chat room, a message box and a message box. Accordingly, there should be a "Send" button. Messages should be sent to all chat participants. The server application should display information about when the user entered and left the chat, the color that the user chose when entering the chat.